Backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
blinker==1.7.0
boto3==1.34.64
botocore==1.34.64
Brotli==1.1.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
deprecation==2.1.0
distro==1.9.0
elevenlabs==0.2.27
executing==2.0.1
filelock==3.13.1
Flask==3.0.2
Flask-Cors==4.0.0
fonttools==4.50.0
fsspec==2024.3.0
future==1.0.0
google-ai-generativelanguage==0.6.2
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.5.2
googleapis-common-protos==1.63.0
gotrue==2.4.2
grpcio==1.63.0
grpcio-status==1.62.2
h11==0.14.0
httpcore==1.0.4
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.21.4
idna==3.7
imageio==2.34.0
imageio-ffmpeg==0.4.9
ipython==8.22.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.3.2
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
moviepy==1.0.3
mutagen==1.47.0
nltk==3.8.1
numpy==1.26.4
openai==1.14.1
opencv-python==4.9.0.80
packaging==24.0
parso==0.8.3
pexpect==4.9.0
Pillow==9.5.0
postgrest==0.16.4
proglog==0.1.10
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
pyloudnorm==0.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytube==15.0.0
PyYAML==6.0.1
realtime==1.0.4
regex==2023.12.25
requests==2.31.0
rsa==4.9
s3transfer==0.10.1
safetensors==0.4.2
scipy==1.12.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.5
supafunc==0.4.5
tokenizers==0.15.2
tqdm==4.66.2
traitlets==5.14.2
transformers==4.38.2
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.1